lab_2/src/main/java/ru/nsu/netesov/lab2/commands/DefineCommand.java
package ru.nsu.netesov.lab2.commands;


import org.apache.log4j.Logger;
import ru.nsu.netesov.lab2.Context;
import ru.nsu.netesov.lab2.exceptions.IncorrectAmountOfArgsException;

import java.util.List;

public class DefineCommand implements Command {
    private final Logger logger = Logger.getLogger(this.getClass());

    @Override
    public void execute(Context context, List<String> args) {
        validate(args);
        String name = args.get(1);
        Double value = Double.parseDouble(args.get(2));
        context.setConstant(name,value);
        logger.info(args.get(0) +  " command finishes work");
    }

    private void validate(List<String> args) {
        if (args.size() != 3) {
            throw new IncorrectAmountOfArgsException("define command needs 2 args");
        }
    }
}
